+++ changelog.md
@@ -55,3 +55,7 @@ Config the System version in **init**.py
 - Initial Development#12: game template finished, and overall improvements on December 20, 2022 -> **0.47**
 
 - Initial Development#12.2: game is finished, and minors overall improvements on December 20, 2022 -> **0.471**
+
+- Initial Development#12.3: fix typing mistake on game on December 20, 2022 -> **0.471b**
+
+- Initial Development#13: cake page done, customized view implementation starts, and overall improvements on December 21, 2022 -> **0.48**